gemini/pyproject.toml
[project]
name = "gemini"
version = "0.1.0"
description = ""
authors = [
    {name = "Mikestriken",email = "michael.john.marais379@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai (>=0.8.4,<0.9.0)"
]

[tool.poetry]
packages = [{include = "gemini", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
